backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
celery[redis]==5.4.*
redis==5.2.*
psycopg2-binary==2.9.*
pydantic==2.9.*
google-genai>=1.0,<2.0
anthropic==0.36.*
stripe==11.*
clerk-backend-api==1.*
svix==1.40.*
weasyprint==62.*
boto3==1.35.*
sentry-sdk[fastapi]==2.*
ddtrace==2.*
python-multipart==0.0.*
python-dotenv==1.*
PyJWT==2.9.*
cryptography==43.*
httpx==0.27.*
